requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
Pillow>=8.3.0
scipy>=1.7.0
torch>=1.9.0
torchvision>=0.10.0
diffusers>=0.21.0
transformers>=4.21.0
opencv-python>=4.5.0
plotly>=5.0.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
tqdm>=4.62.0
jupyter>=1.0.0
ipywidgets>=7.6.0 